feat: Client controller

+++ backend/src/main/java/br/com/quatipunk/repositories/ClientRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import br.com.quatipunk.models.Client;
 
 @Repository
-public interface ClientRepository extends JpaRepository<Client, Long>{
+public interface ClientRepository extends JpaRepository<Client, Integer>{
 }
